Checkpoint: pull existing fakes out into fakes module
There are lots of fakes already in the test code.  This just
extracts some of them.
